+++ src/Models/Referral.php
@@ -23,6 +23,8 @@ use Illuminate\Support\Carbon;
  * @property string|null $coupon_code
  * @property Carbon|null $clicked_at
  * @property Carbon|null $paid_at
+ * @property Carbon|null $refunded_at
+ * @property bool $self_purchase
  * @property Carbon|null $created_at
  */
 class Referral extends Model
@@ -42,6 +44,8 @@ class Referral extends Model
         return [
             'clicked_at' => 'datetime',
             'paid_at' => 'datetime',
+            'refunded_at' => 'datetime',
+            'self_purchase' => 'boolean',
         ];
     }